+++ packages/@headlessui-react/src/components/combobox/combobox.test.tsx
@@ -4699,6 +4699,176 @@ describe('Mouse interactions', () => {
   )
 
   it(
+    'should be possible to open the combobox by focusing the input with immediate mode enabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test" immediate>
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+
+      assertComboboxButton({
+        state: ComboboxState.InvisibleUnmounted,
+        attributes: { id: 'headlessui-combobox-button-2' },
+      })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+
+      // Focus the input
+      await focus(getComboboxInput())
+
+      // Verify it is visible
+      assertComboboxButton({ state: ComboboxState.Visible })
+      assertComboboxList({
+        state: ComboboxState.Visible,
+        attributes: { id: 'headlessui-combobox-options-3' },
+      })
+      assertActiveElement(getComboboxInput())
+      assertComboboxButtonLinkedWithCombobox()
+
+      // Verify we have combobox options
+      let options = getComboboxOptions()
+      expect(options).toHaveLength(3)
+      options.forEach((option) => assertComboboxOption(option))
+    })
+  )
+
+  it(
+    'should not be possible to open the combobox by focusing the input with immediate mode disabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test">
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+
+      assertComboboxButton({
+        state: ComboboxState.InvisibleUnmounted,
+        attributes: { id: 'headlessui-combobox-button-2' },
+      })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+
+      // Focus the input
+      await focus(getComboboxInput())
+
+      // Verify it is invisible
+      assertComboboxButton({
+        state: ComboboxState.InvisibleUnmounted,
+        attributes: { id: 'headlessui-combobox-button-2' },
+      })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+    })
+  )
+
+  it(
+    'should not be possible to open the combobox by focusing the input with immediate mode enabled when button is disabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test" disabled immediate>
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+
+      assertComboboxButton({
+        state: ComboboxState.InvisibleUnmounted,
+        attributes: { id: 'headlessui-combobox-button-2' },
+      })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+
+      // Focus the input
+      await focus(getComboboxInput())
+
+      // Verify it is invisible
+      assertComboboxButton({
+        state: ComboboxState.InvisibleUnmounted,
+        attributes: { id: 'headlessui-combobox-button-2' },
+      })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+    })
+  )
+
+  it(
+    'should be possible to close a combobox on click with immediate mode enabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test" immediate>
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+
+      // Open combobox
+      await click(getComboboxButton())
+
+      // Verify it is visible
+      assertComboboxButton({ state: ComboboxState.Visible })
+
+      // Click to close
+      await click(getComboboxButton())
+
+      // Verify it is closed
+      assertComboboxButton({ state: ComboboxState.InvisibleUnmounted })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+      assertActiveElement(getComboboxInput())
+    })
+  )
+
+  it(
+    'should be possible to close a focused combobox on click with immediate mode enabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test" immediate>
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+      assertComboboxButton({ state: ComboboxState.InvisibleUnmounted })
+
+      // Open combobox by focusing input
+      await focus(getComboboxInput())
+      assertActiveElement(getComboboxInput())
+
+      // Verify it is visible
+      assertComboboxButton({ state: ComboboxState.Visible })
+
+      // Click to close
+      await click(getComboboxButton())
+
+      // Verify it is closed
+      assertComboboxButton({ state: ComboboxState.InvisibleUnmounted })
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
+      assertActiveElement(getComboboxInput())
+    })
+  )
+
+  it(
     'should be possible to open the combobox on click',
     suppressConsoleLogs(async () => {
       render(
@@ -5355,6 +5525,35 @@ describe('Mouse interactions', () => {
 
       // Verify the active option is the previously selected one
       assertActiveComboboxOption(getComboboxOptions()[1])
+    })
+  )
+
+  it(
+    'should be possible to click a combobox option, which closes the combobox with immediate mode enabled',
+    suppressConsoleLogs(async () => {
+      render(
+        <Combobox value="test" immediate>
+          <Combobox.Input onChange={NOOP} />
+          <Combobox.Button>Trigger</Combobox.Button>
+          <Combobox.Options>
+            <Combobox.Option value="a">Option A</Combobox.Option>
+            <Combobox.Option value="b">Option B</Combobox.Option>
+            <Combobox.Option value="c">Option C</Combobox.Option>
+          </Combobox.Options>
+        </Combobox>
+      )
+
+      // Open combobox by focusing input
+      await focus(getComboboxInput())
+      assertActiveElement(getComboboxInput())
+
+      assertComboboxList({ state: ComboboxState.Visible })
+
+      let options = getComboboxOptions()
+
+      // We should be able to click the first option
+      await click(options[1])
+      assertComboboxList({ state: ComboboxState.InvisibleUnmounted })
     })
   )